requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
apscheduler==3.10.4
requests==2.31.0
pytz==2023.3